[handlers] Extract reminder job scheduling
