--- services/api/app/reminders/__init__.py
@@ -1,5 +1,0 @@
-"""Reminder utilities."""
-
-from .common import DefaultJobQueue, schedule_reminder
-
-__all__ = ["DefaultJobQueue", "schedule_reminder"]
